rest: refactor, moved answer and rating resources into its own class
