+++ src/main/java/edu/kit/ipd/crowdcontrol/objectservice/Main.java
@@ -78,6 +78,9 @@ public class Main {
                 new WorkerResource(workerOperations, platformManager),
                 new CalibrationResource(calibrationOperations),
                 new ExperimentResource(experimentOperations, answerRatingOperations, calibrationOperations, tagConstraintsOperations, workerOperations, algorithmsOperations),
-                new AlgorithmResources(algorithmsOperations)).init();
+                new AlgorithmResources(algorithmsOperations),
+                new AnswerRatingResource(experimentOperations, answerRatingOperations, workerOperations)
+        )
+        .init();
     }
 }
